LayoutTests/storage/indexeddb/objectstore-removeobjectstore.html

<html>
<head>
<link rel="stylesheet" href="../../fast/js/resources/js-test-style.css">
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="../../fast/js/resources/js-test-post-function.js"></script>
<script src="resources/shared.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>

description("Test IndexedDB's webkitIDBObjectStore.deleteObjectStore().");
if (window.layoutTestController) 
    layoutTestController.waitUntilDone();

function test()
{
    result = evalAndLog("webkitIndexedDB.open('name')");
    verifyResult(result);
    result.onsuccess = startSetVersion;
    result.onerror = unexpectedErrorCallback;
}

function startSetVersion()
{
    verifySuccessEvent(event);
    db = evalAndLog("db = event.result");

    result = evalAndLog("db.setVersion('new version')");
    verifyResult(result);
    result.onsuccess = deleteExisting;
    result.onerror = unexpectedErrorCallback;
}

function deleteExisting()
{
    verifySuccessEvent(event);
    window.trans = evalAndLog("trans = event.result");
    shouldBeTrue("trans !== null");

    deleteAllObjectStores(db, createObjectStoreAndAddValue);
}

function createObjectStoreAndAddValue()
{
    store = evalAndLog("store = db.createObjectStore('storeName', null)");

    window.index = evalAndLog("store.createIndex('indexName', '')");
    shouldBeTrue("store.indexNames.contains('indexName')");

    result = evalAndLog("store.add('value', 'key')");
    verifyResult(result);
    result.onsuccess = getValue;
    result.onerror = unexpectedErrorCallback;
}

function getValue()
{
    verifySuccessEvent(event);

    transaction = evalAndLog("db.transaction()");
    transaction.onabort = unexpectedErrorCallback;
    var store = evalAndLog("store = transaction.objectStore('storeName')");

    result = evalAndLog("store.get('key')");
    verifyResult(result);
    result.onsuccess = addIndex;
    result.onerror = unexpectedErrorCallback;
}

function addIndex()
{
    verifySuccessEvent(event);
    shouldBeEqualToString("event.result", "value");

    result = evalAndLog("db.setVersion('new version')");
    verifyResult(result);
    result.onsuccess = deleteObjectStore;
    result.onerror = unexpectedErrorCallback;
}

function deleteObjectStore()
{
    verifySuccessEvent(event);
    window.trans = evalAndLog("trans = event.result");
    shouldBeTrue("trans !== null");
    trans.onabort = unexpectedAbortCallback;

    evalAndLog("db.deleteObjectStore('storeName')");
    createObjectStoreAgain();
}

function createObjectStoreAgain()
{
    evalAndLog("db.createObjectStore('storeName', null)");
    getValueAgain();
}

function getValueAgain()
{
    transaction = evalAndLog("db.transaction()");
    transaction.onabort = unexpectedErrorCallback;
    var store = evalAndLog("store = transaction.objectStore('storeName')");

    result = evalAndLog("store.get('key')");
    verifyResult(result);
    result.onsuccess = unexpectedSuccessCallback;
    result.onerror = verifyError;
}

function verifyError()
{
    verifyErrorEvent(event);
    shouldBe("event.code", "webkitIDBDatabaseException.NOT_FOUND_ERR");
    shouldBeFalse("event.source.indexNames.contains('indexName')");

    done();
}

test();

var successfullyParsed = true;
</script>
</body>
</html>
